h3blog_test/requirements.txt
flask
pymysql
pytest
pytest-html
allure-pytest
requests